Move base models to the separate folder
