+++ Pomodoro.Api/ViewModels/SettingsViewModel.cs
@@ -2,6 +2,8 @@
 // Copyright (c) PomodoroGroup_GL_BaseCamp. All rights reserved.
 // </copyright>
 
+using Pomodoro.Api.ViewModels.Base;
+
 namespace Pomodoro.Api.ViewModels
 {
     /// <summary>
